update SEGNET and MaxPooling Layer and CNN evaluation method

+++ code/nnv/engine/nn/layers/Layer.m
@@ -13,8 +13,8 @@ classdef Layer
         % Constructor
         function obj = Layer(W, b, f)
             if size(W, 1) == size(b, 1)
-                obj.W = W;
-                obj.b = b;
+                obj.W = double(W);
+                obj.b = double(b);
                 obj.N = size(W, 1);
             else
                 error('Inconsistent dimensions between Weights matrix and bias vector');
